Minor refactor, export all base classes from an index
